feat auth / crud departemet / crud site

+++ database/migrations/2022_12_07_103033_create_sites_table.php
@@ -19,10 +19,10 @@ return new class extends Migration
             $table->string('name');
             $table->longText('description')->nullable();
             $table->float('price');
-            $table->decimal('latitude', 10, 8)->nullable();
-            $table->decimal('longitude', 11, 8)->nullable();
-            // $table->float('latitude');
-            // $table->float('longitude');
+            $table->float('latitude', 10, 8)->nullable();
+            $table->float('longitude', 11, 8)->nullable();
+            // $table->decimal('latitude', 10, 8)->nullable();
+            // $table->decimal('longitude', 11, 8)->nullable();
             $table->timestamps();
         });
     }
